エラー画面関連リファクタリング
* views/bad_request.html.erbを追加
* 400エラー画面の変更関連を追加 rescue_from ActionController::ParameterMissing, with: :rescue400
*　concerns/error_handlersにerrorモジュールを作成
* application_controller.rbリファクタリング実施
 - include ErrorHandlers if Rails.env.productionプロダクション環境のみにエラーが画面が適用されるように追加
* 各所コメント追加
