+++ app/views/errors/bad_request.html.erb
@@ -1,0 +1,4 @@
+<div id="error">
+  <h1>400 Bad Request</h1>
+  <p>不正な要求だよ( *´艸｀)</p>
+</div>
